Rename data package to infrastructure
