+++ lib/infrastructure/models/checklist_model.dart
