+++ lib/infrastructure/models/comment_model.dart
@@ -1,4 +1,4 @@
-import 'package:aspdm_project/data/models/user_model.dart';
+import 'package:aspdm_project/infrastructure/models/user_model.dart';
 import 'package:aspdm_project/domain/entities/comment.dart';
 import 'package:aspdm_project/domain/values/task_values.dart';
 import 'package:aspdm_project/domain/values/unique_id.dart';
