+++ lib/infrastructure/models/label_model.dart
@@ -1,5 +1,5 @@
 import 'package:aspdm_project/domain/entities/label.dart';
-import 'package:aspdm_project/data/color_parser.dart';
+import 'package:aspdm_project/infrastructure/color_parser.dart';
 import 'package:aspdm_project/domain/values/unique_id.dart';
 import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
