+++ lib/locator.dart
@@ -1,14 +1,14 @@
-import 'package:aspdm_project/data/repositories/archive_repository_impl.dart';
-import 'package:aspdm_project/data/repositories/auth_repository_impl.dart';
-import 'package:aspdm_project/data/repositories/home_repository_impl.dart';
-import 'package:aspdm_project/data/repositories/task_repository_impl.dart';
+import 'package:aspdm_project/infrastructure/repositories/archive_repository_impl.dart';
+import 'package:aspdm_project/infrastructure/repositories/auth_repository_impl.dart';
+import 'package:aspdm_project/infrastructure/repositories/home_repository_impl.dart';
+import 'package:aspdm_project/infrastructure/repositories/task_repository_impl.dart';
 import 'package:aspdm_project/domain/repositories/archive_repository.dart';
 import 'package:aspdm_project/domain/repositories/auth_repository.dart';
 import 'package:aspdm_project/domain/repositories/home_repository.dart';
 import 'package:aspdm_project/domain/repositories/task_repository.dart';
 import 'package:aspdm_project/services/app_info_service.dart';
 import 'package:aspdm_project/services/connectivity_service.dart';
-import 'package:aspdm_project/data/datasources/remote_data_source.dart';
+import 'package:aspdm_project/infrastructure/datasources/remote_data_source.dart';
 import 'package:aspdm_project/services/log_service.dart';
 import 'package:aspdm_project/services/navigation_service.dart';
 import 'package:aspdm_project/services/preference_service.dart';
