+++ test/infrastructure/repositories/archive_repository_test.dart
@@ -1,8 +1,8 @@
 import 'package:aspdm_project/core/either.dart';
-import 'package:aspdm_project/data/datasources/remote_data_source.dart';
-import 'package:aspdm_project/data/models/task_model.dart';
-import 'package:aspdm_project/data/models/user_model.dart';
-import 'package:aspdm_project/data/repositories/archive_repository_impl.dart';
+import 'package:aspdm_project/infrastructure/datasources/remote_data_source.dart';
+import 'package:aspdm_project/infrastructure/models/task_model.dart';
+import 'package:aspdm_project/infrastructure/models/user_model.dart';
+import 'package:aspdm_project/infrastructure/repositories/archive_repository_impl.dart';
 import 'package:aspdm_project/domain/failures/server_failure.dart';
 import 'package:aspdm_project/domain/repositories/archive_repository.dart';
 import 'package:flutter_test/flutter_test.dart';
