+++ test/infrastructure/repositories/auth_repository_test.dart
@@ -1,7 +1,7 @@
 import 'package:aspdm_project/core/either.dart';
-import 'package:aspdm_project/data/datasources/remote_data_source.dart';
-import 'package:aspdm_project/data/models/user_model.dart';
-import 'package:aspdm_project/data/repositories/auth_repository_impl.dart';
+import 'package:aspdm_project/infrastructure/datasources/remote_data_source.dart';
+import 'package:aspdm_project/infrastructure/models/user_model.dart';
+import 'package:aspdm_project/infrastructure/repositories/auth_repository_impl.dart';
 import 'package:aspdm_project/domain/entities/user.dart';
 import 'package:aspdm_project/domain/repositories/auth_repository.dart';
 import 'package:aspdm_project/domain/values/unique_id.dart';
